testing/tests/automated/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
pytest-html>=4.1.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0
pytest-ordering>=0.6
pytest-repeat>=0.9.3
allure-pytest>=2.13.0
